+++ src/Curl/CaseInsensitiveArray.php
@@ -42,12 +42,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @param  mixed[]              $initial (optional) Existing array to convert.
      * @return CaseInsensitiveArray
      */
-    // TODO: Use a nullable type declaration when supported versions >= PHP 7.1.
-    //   Trying to use the nullable type declaration on PHP 7.0:
-    //     public function __construct(?array $initial = null)
-    //   results in:
-    //     ParseError: syntax error, unexpected '?', expecting variable (T_VARIABLE)
-    public function __construct($initial = null)
+    public function __construct(?array $initial = null)
     {
         if ($initial !== null) {
             foreach ($initial as $key => $value) {
@@ -93,7 +88,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     #[\ReturnTypeWillChange]
     public function offsetExists($offset)
     {
-        return (bool) array_key_exists(strtolower($offset), $this->data);
+        return array_key_exists(strtolower($offset), $this->data);
     }
 
     /**
@@ -134,20 +129,18 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     /**
      * Count
      *
-     * @param void
      * @return int The number of elements stored in the array.
      * @see https://secure.php.net/manual/en/countable.count.php
      */
     #[\ReturnTypeWillChange]
     public function count()
     {
-        return (int) count($this->data);
+        return count($this->data);
     }
 
     /**
      * Current
      *
-     * @param void
      * @return mixed Data at the current position.
      * @see https://secure.php.net/manual/en/iterator.current.php
      */
@@ -160,7 +153,6 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     /**
      * Next
      *
-     * @param void
      * @return void
      * @see https://secure.php.net/manual/en/iterator.next.php
      */
@@ -173,7 +165,6 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     /**
      * Key
      *
-     * @param void
      * @return mixed Case-sensitive key at current position.
      * @see https://secure.php.net/manual/en/iterator.key.php
      */
@@ -193,13 +184,12 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     #[\ReturnTypeWillChange]
     public function valid()
     {
-        return (bool) (key($this->data) !== null);
+        return (key($this->data) !== null);
     }
 
     /**
      * Rewind
      *
-     * @param void
      * @return void
      * @see https://secure.php.net/manual/en/iterator.rewind.php
      */