+++ src/Curl/MultiCurl.php
@@ -577,6 +577,7 @@ class MultiCurl extends BaseCurl
         $unit = strtolower($matches['3']);
 
         // Convert interval to seconds based on unit.
+        $interval_seconds = '';
         if ($unit === 's') {
             $interval_seconds = $interval * 1;
         } elseif ($unit === 'm') {
@@ -919,7 +920,7 @@ class MultiCurl extends BaseCurl
      */
     private function waitUntilRequestQuotaAvailable()
     {
-        $sleep_until = $this->currentStartTime + $this->intervalSeconds;
+        $sleep_until = (float)($this->currentStartTime + $this->intervalSeconds);
         $sleep_seconds = $sleep_until - microtime(true);
 
         // Avoid using time_sleep_until() as it appears to be less precise and not sleep long enough.
@@ -935,5 +936,10 @@ class MultiCurl extends BaseCurl
         }
 
         $this->currentRequestCount = 0;
+    }
+
+    public function getActiveCurls()
+    {
+        return $this->activeCurls;
     }
 }